pyproject.toml
[tool.poetry]
name = "jamsapi"
version = "0.1.0"
description = "A basic API for JAMS. This is a work in progress."
authors = ["Arpan Pandey <arpanpandey.aps@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.101.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pg8000 = "^1.30.1"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
